extract qr code list
