--- RN/app/(tabs)/types.ts
@@ -1,4 +1,0 @@
-export type HomeListItemType = {
-  title: string;
-  url: string;
-};
